Move ResponseGenerationException from openai utils to llm utils
