+++ bobweb/bob/test_openai_api_utils.py
@@ -8,10 +8,11 @@ from unittest import mock
 
 import bobweb.bob.config
 from bobweb.bob import main, openai_api_utils, database, command_gpt, config
-from bobweb.bob.openai_api_utils import ResponseGenerationException, \
-    remove_openai_related_command_text_and_extra_info, ChatMessage, \
-    msg_serializer_for_text_models, msg_serializer_for_vision_models, GptModel, \
+from bobweb.bob.openai_api_utils import remove_openai_related_command_text_and_extra_info, \
+    ChatMessage, msg_serializer_for_text_models, \
+    msg_serializer_for_vision_models, GptModel, \
     determine_suitable_model_for_version_based_on_message_history, gpt_4o
+from bobweb.bob.litellm_utils import ResponseGenerationException
 from bobweb.bob.telethon_service import ContentOrigin
 from bobweb.bob.test_command_gpt import MockLiteLLMResponseObject
 from bobweb.bob.tests_mocks_v2 import init_chat_user, MockChat, MockUser
